Fix eeprom defaults.

+++ ESP32LapTimer/settings_eeprom.h
@@ -22,7 +22,7 @@ struct EepromSettingsStruct {
   uint8_t RXBand[MAX_NUM_PILOTS];
   uint8_t RXChannel[MAX_NUM_PILOTS];
   uint16_t RXfrequencies[MAX_NUM_PILOTS];
-  int RSSIthresholds[MAX_NUM_PILOTS];
+  uint16_t RSSIthresholds[MAX_NUM_PILOTS];
   RXADCfilter_ RXADCfilter;
   ADCVBATmode_ ADCVBATmode;
   float VBATcalibration;
